% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_with_examples.R
\name{f_with_examples}
\alias{f_with_examples}
\title{Says hello, has examples}
\usage{
f_with_examples()
}
\description{
Says hello, has examples
}
\examples{
  f_with_examples()
}
\author{
Richel Bilderbeej
}

